requirements.txt
flask
fpdf
google-cloud-speech
python-dotenv
uvicorn
